requirements.txt
imageio
numpy
scipy
tensorflow
scikit-learn
pandas
hdf5storage
h5py
matplotlib
steganogan
python-magic
efficientnet
Pillow